AI_OCR_Module2/requirements.txt
opencv-python
pytesseract
pillow
pandas
